Package and Notebook working with imports
