--- distributed-learning-contributivity/constants.py
@@ -1,20 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Declaration of constants.
-"""
-
-# ML constants
-DEFAULT_BATCH_SIZE = 256
-MAX_BATCH_SIZE = 2**20
-DEFAULT_GRADIENT_UPDATES_PER_PASS_COUNT = 8
-PATIENCE = 10  # patience for early stopping
-
-# GPU
-GPU_MEMORY_LIMIT_MB = 4096
-
-# Logging
-INFO_LOGGING_FILE_NAME = "info.log"
-DEBUG_LOGGING_FILE_NAME = "debug.log"
-
-# Contributivity methods names
-# TODO
